refactoring: 중복 타입 삭제
