+++ client/src/api/order/postExcelToDb.ts
@@ -1,18 +1,8 @@
 import { API_END_POINT } from "@/constant";
 import { httpClient } from "@/shared/axios";
+import type { SimpleExcelUploadRequest, ExcelUploadResponse } from "@/shared/types";
 
-interface ExcelUploadRequest {
-  template_code: string;
-  file: File;
-}
-
-interface ExcelUploadResponse {
-  file_url: string;
-  object_name: string;
-  template_code: string;
-}
-
-export const postExcelToDb = async (data: ExcelUploadRequest) => {
+export const postExcelToDb = async (data: SimpleExcelUploadRequest) => {
   const formData = new FormData();
 
   formData.append('template_code', data.template_code);
