+++ client/src/shared/types/order.types.ts
@@ -73,7 +73,6 @@ export interface OrderRegisterData {
   orderData: any[];
 }
 
-
 export interface OrderRegisterFormData {
   selectedTemplate: string;
   orderData: any[];
@@ -166,24 +165,45 @@ export interface ExcelUploadResponse {
   template_code: string;
 }
 
-export interface ExcelUploadFilters {
-  order_date_from: string;
-  order_date_to: string;
-}
-
-export interface ExcelUploadRequestData {
-  template_code: string;
-  created_by: string;
-  filters: ExcelUploadFilters;
-  source_table: string;
-}
-
 export interface ExcelUploadFormData {
   template_code: string;
   order_date_from: string;
   order_date_to: string;
   source_table: string;
   file: File | null;
+}
+
+// Simple Excel upload request for direct file uploads
+export interface SimpleExcelUploadRequest {
+  template_code: string;
+  file: File;
+}
+
+export interface DownFormOrderResponse {
+  item: any;
+  status: string;
+  message: string;
+}
+
+export interface PaginationResponse {
+  total: number;
+  page: number;
+  page_size: number;
+  items: DownFormOrderResponse[];
+}
+
+export interface GetDownFormOrdersPaginationParams {
+  page?: number;
+  page_size?: number;
+  template_code?: string;
+}
+
+export interface UseDownFormOrderPaginationParams extends GetDownFormOrdersPaginationParams {
+  enabled?: boolean;
+}
+
+export interface UseOrderDataParams {
+  page: number;
 }
 
 export type OrderTab = "registration" | "bulk-registration";
